Improve value/state flow for TextInput and UnitInput
As well as Select, Slider, and SearchInput
