+++ packages/components/src/ColorPicker/ColorPickerInputHex.js
@@ -6,7 +6,6 @@ import React from 'react';
 import { FormGroup } from '../FormGroup';
 import { Text } from '../Text';
 import { TextInput } from '../TextInput';
-import { useNumberKeyboardHandlers } from '../TextInput/useTextInputNumberState';
 import { useBaseDragHandlers } from '../TextInput/useTextInputState.utils';
 import { useColorPickerContext } from './ColorPicker.Context';
 
@@ -30,12 +29,6 @@ export const ColorInputHex = React.memo(
 			dragAxis: 'y',
 		});
 
-		const keyboardHandlers = useNumberKeyboardHandlers({
-			increment,
-			decrement,
-			stopIfEventDefaultPrevented: false,
-		});
-
 		const handleOnChange = React.useCallback(
 			(next) => {
 				setValue(ui.color(next).toHexString());
@@ -52,7 +45,6 @@ export const ColorInputHex = React.memo(
 				<TextInput
 					{...otherProps}
 					{...dragHandlers}
-					{...keyboardHandlers}
 					onChange={handleOnChange}
 					prefix={<Text variant="muted">#</Text>}
 					validate={handleOnValidate}
