+++ packages/components/src/Select/useSelect.js
@@ -1,6 +1,6 @@
 import { useContextSystem } from '@wp-g2/context';
 import { css, cx, ui } from '@wp-g2/styles';
-import { is, noop, useControlledState, useResizeAware } from '@wp-g2/utils';
+import { is, noop, useControlledValue, useResizeAware } from '@wp-g2/utils';
 import React from 'react';
 import { useCallback, useRef, useState } from 'react';
 
@@ -29,7 +29,7 @@ export function useSelect(props) {
 		isSubtle,
 		multiple,
 		onBlur = noop,
-		onChange = noop,
+		onChange: onChangeProp = noop,
 		onFocus = noop,
 		options = [],
 		placeholder: placeholderProp,
@@ -40,9 +40,12 @@ export function useSelect(props) {
 		...otherProps
 	} = useContextSystem(props, 'Select');
 
-	const [value, setValue] = useControlledState(valueProp, {
-		initial: multiple ? [] : defaultValue,
+	const [value, onChange] = useControlledValue({
+		value: valueProp,
+		onChange: onChangeProp,
+		defaultValue: multiple ? [] : defaultValue,
 	});
+
 	const [isFocused, setIsFocused] = useState(isFocusedProp);
 	const [resizer, sizes] = useResizeAware();
 	/** @type {import('react').Ref<HTMLSelectElement | undefined>} */
@@ -100,10 +103,10 @@ export function useSelect(props) {
 				next = event.target.value;
 			}
 
-			setValue(next);
+			console.log(next);
 			onChange(next, { event });
 		},
-		[multiple, setValue, onChange],
+		[multiple, onChange],
 	);
 
 	const shouldRenderPlaceholder =
