+++ packages/components/src/Slider/useSlider.js
@@ -6,7 +6,7 @@ import {
 	is,
 	noop,
 	parseUnitValue,
-	useControlledState,
+	useControlledValue,
 } from '@wp-g2/utils';
 import { useCallback, useMemo, useState } from 'react';
 
@@ -29,7 +29,7 @@ export function useSlider(props) {
 		defaultValue,
 		error,
 		onBlur = noop,
-		onChange = noop,
+		onChange: onChangeProp = noop,
 		onFocus = noop,
 		id: idProp,
 		isFocused: isFocusedProp = false,
@@ -40,10 +40,14 @@ export function useSlider(props) {
 		value: valueProp,
 		...otherProps
 	} = useContextSystem(props, 'Slider');
-	const [initialValue, initialUnit] = parseUnitValue(valueProp);
-	const [value, setValue] = useControlledState(initialValue, {
-		initial: defaultValue || 50,
+
+	const [_value, onChange] = useControlledValue({
+		value: valueProp,
+		onChange: onChangeProp,
+		defaultValue: defaultValue || 50,
 	});
+	const [value, initialUnit] = parseUnitValue(_value);
+
 	const id = useFormGroupContextId(idProp);
 	const [isFocused, setIsFocused] = useState(isFocusedProp);
 
@@ -66,10 +70,9 @@ export function useSlider(props) {
 				next = createUnitValue(nextValue, initialUnit);
 			}
 
-			setValue(nextValue);
 			onChange(next, { event });
 		},
-		[onChange, setValue, initialUnit],
+		[onChange, initialUnit],
 	);
 
 	const handleOnFocus = useCallback(
