+++ packages/components/src/TextInput/useTextInput.js
@@ -1,6 +1,5 @@
 import { useContextSystem } from '@wp-g2/context';
 import { cx } from '@wp-g2/styles';
-import { noop } from '@wp-g2/utils';
 import { useCallback, useMemo } from 'react';
 import TextareaAutosize from 'react-textarea-autosize';
 
@@ -8,11 +7,8 @@ import { useBaseField } from '../BaseField';
 import { useFormGroupContextId } from '../FormGroup';
 import * as styles from './TextInput.styles';
 import { useTextInputState } from './useTextInputState';
-import { useDragHandlers } from './useTextInputState.utils';
 
-const useRootEventHandlers = ({ decrement, increment, inputRef, store }) => {
-	const dragHandlers = useDragHandlers({ store, increment, decrement });
-
+const useRootEventHandlers = ({ dragHandlers, inputRef }) => {
 	const handleOnClick = useCallback(() => {
 		inputRef.current.focus();
 	}, [inputRef]);
@@ -35,7 +31,6 @@ const useRootEventHandlers = ({ decrement, increment, inputRef, store }) => {
 export function useTextInput(props) {
 	const combinedProps = useContextSystem(props, 'TextInput');
 	const {
-		__debugger = false,
 		align,
 		arrows = true,
 		className,
@@ -46,16 +41,15 @@ export function useTextInput(props) {
 		format,
 		gap = 2.5,
 		id: idProp,
-		isInline = false,
-		isFocused: isFocusedProp,
 		isCommitOnBlurOrEnter = true,
+		isFocused: isFocusedProp,
+		isInline = false,
 		isResizable = false,
 		isShiftStepEnabled = true,
 		justify,
 		max,
 		min,
 		multiline = false,
-		onValueChange = noop,
 		prefix,
 		shiftStep = 10,
 		size = 'medium',
@@ -71,20 +65,23 @@ export function useTextInput(props) {
 
 	const {
 		decrement,
+		dragHandlers,
 		increment,
-		store,
+		inputRef,
+		isFocused,
+		isInputTypeNumeric,
+		isTypeNumeric,
+		value,
 		...textInputState
 	} = useTextInputState({
 		...otherProps,
-		__debugger,
 		format,
-		initialValue: defaultValue,
+		defaultValue,
 		isCommitOnBlurOrEnter,
 		isFocused: isFocusedProp,
 		isShiftStepEnabled,
 		max,
 		min,
-		onValueChange,
 		shiftStep,
 		step,
 		value: valueProp,
@@ -92,13 +89,9 @@ export function useTextInput(props) {
 		type,
 	});
 
-	const { inputRef, isFocused, isTypeNumeric, value } = store();
-
 	const rootEventHandlers = useRootEventHandlers({
 		inputRef,
-		decrement,
-		increment,
-		store,
+		dragHandlers,
 	});
 
 	const baseFieldProps = useBaseField({
@@ -151,18 +144,18 @@ export function useTextInput(props) {
 	return {
 		...baseFieldProps,
 		...rootEventHandlers,
-		__store: store,
 		arrows,
 		className: classes,
-		dragAxis,
 		decrement,
-		increment,
+		disabled,
+		dragAxis,
+		dragHandlers,
 		format,
+		increment,
 		inputProps,
 		inputRef,
 		isTypeNumeric,
 		prefix,
 		suffix,
-		disabled,
 	};
 }
