+++ packages/utils/src/hooks/index.js
@@ -1,5 +1,6 @@
 export * from './use-clipboard';
 export * from './use-controlled-state';
+export * from './use-controlled-value';
 export * from './use-isomorphic-layout-effect';
 export * from './use-local-state';
 export * from './use-media-query';
